create_trello_card/requirements.txt
py-trello==0.6.1
